+++ src/main/java/com/jockey/community/dto/QuestionDTO.java
@@ -5,8 +5,8 @@ import lombok.Data;
 
 @Data
 public class QuestionDTO {
-    private int id;
-    private int creator;
+    private Long id;
+    private Long creator;
     private String title;
     private String description;
     private Long gmtCreate;
