+++ src/main/resources/db/migration/V3__Create_question_table.sql
@@ -1,10 +1,10 @@
 CREATE TABLE question (
-  id int NOT NULL AUTO_INCREMENT primary key,
+  id bigint NOT NULL AUTO_INCREMENT primary key,
   title varchar(50),
   description text,
   gmt_create bigint,
   gmt_modified bigint,
-  creator int,
+  creator bigint,
   comment_count int DEFAULT 0,
   view_count int DEFAULT 0,
   like_count int DEFAULT 0,
